Add Basic API for FluxCD Application
