+++ pkg/kapis/gitops/v1alpha1/argocd/route_test.go
@@ -1,30 +1,12 @@
-// Copyright 2022 KubeSphere Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
 package argocd
 
 import (
-	"bytes"
 	"fmt"
 	"github.com/emicklei/go-restful"
 	"github.com/stretchr/testify/assert"
 	"io"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"kubesphere.io/devops/pkg/api"
 	"kubesphere.io/devops/pkg/api/gitops/v1alpha1"
 	"kubesphere.io/devops/pkg/apiserver/runtime"
 	"kubesphere.io/devops/pkg/config"
@@ -33,7 +15,6 @@ import (
 	"net/http/httptest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/yaml"
 	"testing"
 )
 
@@ -128,168 +109,6 @@ func TestAPIs(t *testing.T) {
 		k8sclient    client.Client
 		verify       func(t *testing.T, body []byte)
 	}{{
-		name: "get an empty list of the applications",
-		request: request{
-			method: http.MethodGet,
-			uri:    "/namespaces/fake/applications",
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &api.ListResult{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			assert.Equal(t, 0, len(list.Items))
-		},
-	}, {
-		name: "get a normal list of the applications",
-		request: request{
-			method: http.MethodGet,
-			uri:    "/namespaces/ns/applications",
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema, app.DeepCopy()),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &api.ListResult{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			assert.Equal(t, 1, len(list.Items))
-			assert.Nil(t, err)
-		},
-	}, {
-		name: "get a normal application",
-		request: request{
-			method: http.MethodGet,
-			uri:    "/namespaces/ns/applications/app",
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema, app.DeepCopy()),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &unstructured.Unstructured{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			name, _, err := unstructured.NestedString(list.Object, "metadata", "name")
-			assert.Equal(t, "app", name)
-			assert.Nil(t, err)
-		},
-	}, {
-		name: "delete an application",
-		request: request{
-			method: http.MethodDelete,
-			uri:    "/namespaces/ns/applications/app",
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema, app.DeepCopy()),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &unstructured.Unstructured{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			name, _, err := unstructured.NestedString(list.Object, "metadata", "name")
-			assert.Equal(t, "app", name)
-			assert.Nil(t, err)
-		},
-	}, {
-		name: "delete an application by cascade",
-		request: request{
-			method: http.MethodDelete,
-			uri:    "/namespaces/ns/applications/app?cascade=true",
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema, app.DeepCopy()),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &unstructured.Unstructured{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			name, _, err := unstructured.NestedString(list.Object, "metadata", "name")
-			assert.Equal(t, "app", name)
-			finalizers, _, err := unstructured.NestedSlice(list.Object, "metadata", "finalizers")
-			assert.Equal(t, []interface{}{"resources-finalizer.argocd.argoproj.io"}, finalizers)
-			assert.Nil(t, err)
-		},
-	}, {
-		name: "create an application",
-		request: request{
-			method: http.MethodPost,
-			uri:    "/namespaces/ns/applications",
-			body: func() io.Reader {
-				return bytes.NewBuffer([]byte(`{
-  "apiVersion": "devops.kubesphere.io/v1alpha1",
-  "kind": "Application",
-  "metadata": {
-    "name": "fake"
-  },
-  "spec": {
-    "argoApp": {
-      "spec": {
-        "project": "default"
-      }
-    }
-  }
-}`))
-			},
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &unstructured.Unstructured{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			name, _, err := unstructured.NestedString(list.Object, "metadata", "name")
-			assert.Equal(t, "fake", name)
-			assert.Nil(t, err)
-		},
-	}, {
-		name: "create an application, invalid payload",
-		request: request{
-			method: http.MethodPost,
-			uri:    "/namespaces/ns/applications",
-			body: func() io.Reader {
-				return bytes.NewBuffer([]byte(`fake`))
-			},
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema),
-		responseCode: http.StatusInternalServerError,
-	}, {
-		name: "update an application",
-		request: request{
-			method: http.MethodPut,
-			uri:    "/namespaces/ns/applications/app",
-			body: func() io.Reader {
-				return bytes.NewBuffer([]byte(`{
-  "apiVersion": "devops.kubesphere.io/v1alpha1",
-  "kind": "Application",
-  "metadata": {
-    "name": "app",
-    "namespace": "ns"
-  },
-  "spec": {
-    "argoApp": {
-      "spec": {
-        "project": "good"
-      }
-    }
-  }
-}`))
-			},
-		},
-		k8sclient:    fake.NewFakeClientWithScheme(schema, app.DeepCopy()),
-		responseCode: http.StatusOK,
-		verify: func(t *testing.T, body []byte) {
-			list := &unstructured.Unstructured{}
-			err := yaml.Unmarshal(body, list)
-			assert.Nil(t, err)
-
-			project, _, err := unstructured.NestedString(list.Object, "spec", "argoApp", "spec", "project")
-			assert.Equal(t, "good", project)
-			assert.Nil(t, err)
-		},
-	}, {
 		name: "get clusters, no expected data",
 		request: request{
 			method: http.MethodGet,
