+++ pkg/kapis/gitops/v1alpha1/util.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-package argocd
+package v1alpha1
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
