+++ pkg/kapis/gitops/v1alpha1/util_test.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-package argocd
+package v1alpha1
 
 import (
 	"github.com/stretchr/testify/assert"
@@ -23,7 +23,7 @@ import (
 	"testing"
 )
 
-func Test_toObjects(t *testing.T) {
+func Test_ToObjects(t *testing.T) {
 	createApplication := func(name, namespaces string) *v1alpha1.Application {
 		return &v1alpha1.Application{
 			ObjectMeta: metav1.ObjectMeta{
